Pull livefs launchpad mapping in from 'production' branch.
While it's useful to be able to separate the configuration from the code so
that third parties can deploy without having to fork the code, in practice
since the production branch also contains passwords, it is not public and
therefore leaves downstreams with no examples to follow.  It is also
confusing to people contributing new images to the branch because they don't
have visibility on the other branch that has to be updated at the same time.

Try a different approach, by moving the public information (starting with
the livefs-launchpad file) into etc/ while allowing downstreams to derive by
dropping their own file in production/.