uc20: build grade-dangerous images too
Currently, test-lab can only use `grade: dangerous` UC20 images for
x86 provisioning, due to requirement for unrestricted access to
cloud-init which is currently not available for `grade: signed`
models. Dangerous is not a channel, per-se, but a model that allows
overriding channels at build time, but for the purpose of
cdimage/livecd-rootfs we treat it as just another risk level.