sign: Work with gpg2
gpg2 doesn't use secring.gpg. On first launch it will import the keys from a
pre-existing secring.gpg to a new directory called private-keys-v1.d. This
doesn't respect the --secret-keyring parameter, though. But gpg (both xenial's
and focal's) have a `--homedir` option which allows you to override the
location for all files. This one does work. Switch to it, and remove the direct
setting of the various keyring paths.