default-arches: Disable riscv64
This has shaken out build failures which the team is now working on. In the
meantime, we're prevented from getting daily-preinstalled builds for hirsute
server, and for focal subscribers are getting emailed daily about failures.
Once people are actively working on getting these builds up and running, this
commit can be reverted.