mirror: Add a flag file etc/STOP_SYNC_MIRRORS to stop syncing
Our current method of `chmod -x bin/sync-mirrors` doesn't work. In particular,
the mirrors are triggered whenever any daily build is completed. This means
that in progress releases are highly vulnerable to leaking early. In the past
we have asked Canonical IS to stop rsync to prevent this from happening. But we
can take care of it on our side by avoiding ever asking the frontends to pull
from us.

A better API where we can ask frontends to only sync *certain* images - the
daily which just finished - would be preferable.