Collapse references to obsolete series
We don't need to include complex if/else logic for calculating the target
size of images for series we will never build again.  Continue to allow for
the possibility that we will rebuild images for bionic just in case, but
drop references for any ESM releases and EOLEd non-LTS releases.  The worst
consequence of getting this wrong is that an extraordinary image build will
be bigger than the original and we won't get warnings about it, which is not
really a serious problem.