Adjust ubuntu-server jobs to all start at roughly the same time.
riscv64 are our longest-running builds, and currently there is no idle time
when it's possible for us to sync the ftp mirror (which policy currently
requires that all builds be idle).  These are highly parallelizable on the
buildd side, so move them closer together in time to increase chances of an
idle window.