Don't block ftp mirror syncing while other images are building
We have prevented rsync from running on the ftp mirror while other image
builds are in progress due to concerns that the rsync could delete a file
that the other build needs, before it has a chance to access it.  This is no
longer a concern because the file retention policy on the rsync source
(ftpmaster.internal) now guarantees us files are referenced via by-hash
index files for 2 days, so will definitely not be garbage-collected in the
time it takes a single image build to finish.

The mirror updates are currently failing because we have enough image builds
happening that cdimage-master is never idle, resulting in buggy image builds
with out-of-date package pools.  So we should just drop the lock handling.

This also drops the multipidfile code, which nothing else in the codebase
was referencing.

Ubuntu-Bug: https://bugs.launchpad.net/ubuntu-cdimage/+bug/1976146