rsync now goes to ubuntu-archive-toolbox.internal, not archive-team.internal.
archive-team.internal is snakefruit which is being turned off.  The
archive-team.internal name is used for at least two separate things: it's
where ancientminister.internal pushes its image build logs via rsync, and
it's where livecd-rootfs pulls its seed information via http.

As of right now, the snakefruit replacement ubuntu-archive-toolbox.internal
is the right place to send logs, but that machine is not currently serving
content over http; that functionality has been split to a different machine.

So send to ubuntu-archive-toolbox.internal by name rather than by the alias
which may or not point to this machine in the end.