don't use complicated subarch name mapping between cdimage and livecd-rootfs
riscv64+unmatched images for beta didn't build right, the correct logic
wasn't applied for the boot partitions on the image.  Trying to fix this in
livecd-rootfs broke it further, because the subarch being passed to
livecd-rootfs by ubuntu-cdimage is 'sifive_hifive_unmatched_fu740' even
though all the publishing artifacts end up as '+unmatched'.

This is unnecessarily confusing.  We should just use the same subarch name
throughout.

Leaving this in place for jammy, as the jammy livecd-rootfs has not been
changed.