Drop support for trusty and xenial (except xenial for core16).
debian-cd has dropped support for building any images older than jammy.
Here we follow partway, dropping support for images older than bionic.

This also drops support for powerpc, which was last available in xenial;
and amd64+mac, which was last available in trusty.

It furthermore drops support for metalinks, which were only used in xenial
and earlier; and does not expect wubi as a download artifact.