write a per-arch list of packages to include in the pool
Currently the way cdimage tells debian-cd which packages to include in
the pool via this absurd mechanism involving the C preprocessor on the
debian-cd side. But cdimage knows exactly which packages to include, so
write a simple file with the list of packages we want (alongside the
current silliness, for now).