Diable proposed for the daily noble build
Steve said that building daily images with -proposed was to support
testing version of the installer from -proposed. However, given the move
to subiquity and ubuntu-desktop-installer which are both snap based this
no longer is needed. Additionally, packages can end up stuck in
-proposed for a variety of reasons which could impact the quality of the
daily image. Creating daily images with -proposed should be the
exception not the norm.