Explicitly mark flavors that use calamares as building with PROPOSED=1
Outside of milestone prep, we build daily images for stable series using
packages from the proposed pocket in order to be able to test installer
updates.  It is no longer the common case that images use installers from
debs, so building from proposed is counterproductive.  Explicitly set
PROPOSED=1 for the flavors that do use an installer as a deb.