report if just directly downloading all livefs outputs would differ
There is some considerable hair in ubuntu-cdimage for deciding which
artefacts to download from a livefs build in Launchpad and the paths to
save them under.

However, I *think* this is all unnecessary for all the builds we
actually make today: we could just list all the build outputs,
mechanically translate the livecd.${PROJECT}. prefix to
${ARCH}+${SUBARCH}. and download them to this path.

But because the hair is so hairy it's hard to be sure. So this branch
adds a check of the paths that were downloaded, builds a lists of the
paths that would have been downloaded by the above scheme, compares them
and dumps a diff into the log if they differ. Then we can deploy this,
check the logs for a few days and hopefully delete all the complexity.
Which will make deleting further complexity (i.e. support for
non-Launchpad livefs builders) easier because a lot of fragile tests
will have been deleted.

I actually know of one project that does not download all the artefacts,
which is the ubuntu-mini-iso project (the livefs builds a rootfs tarball
that is not downloaded) and this lets me show off what the output looks
like:

DOWNLOADED PATHS DID NOT MATCH DIRECT DOWNLOAD!
--- actual
+++ direct
@@ -1,2 +1,3 @@
 amd64.iso
 amd64.manifest
+amd64.rootfs.tar.gz