ignore some trivial differences when comparing direct downloads
My previous commit tried to check whether the complicated selection and
renaming of livefs artefacts that happens in livefs.py currently is
still needed by comparing the downloaded artefacts with what would
happen with a more direct download. Running this in prod for 24 hours or
so reveals some trivial differences:

1) some livefs builds produce ".manifest.full" files which are not
   downloaded
2) all squashfs files get signed and so there are .gpg files my code was
   not expecting.

Both of these are very boring differences so this change will ignore
them, and hopefully running *this* version in production for a few days
will let me know if I can delete lots of code.