maintain explicit mapping of arch to livefs build
The previous code did this game of starting a build for each arch (which
could be cpuarch+subarch), just storing a list of builds, and then
trying to figure out which build was started for a given arch. Stop
doing that and store a mapping from arch to build instead.

This also lets me apply a couple more tweaks to the code that derives
which files to download (and where to download them to) from the build's
file urls to hopefully match what the current complicated code does.