add retries to osextras.fetch
A lack of retries here caused the noble image dated 2025-02-10 to be
broken:
```
2025-02-10 07:59:44 URL:https://launchpadlibrarian.net/775682399/livecd.ubuntu.minimal.standard.it.squashfs [2818048/2818048] -> "/srv/cdimage.ubuntu.com/scratch/ubuntu/noble/daily-live/live/amd64.minimal.standard.it.squashfs" [1]
https://launchpadlibrarian.net/775682403/livecd.ubuntu.minimal.standard.live.squashfs:
2025-02-10 08:01:45 ERROR 503: Service Unavailable.
```

Which caused bug LP: #2031899 to resurface. This commit should alleviate
said bug and hopefully cause it to not occur again, apart from the in
the instance of a widespread launchpad outage, in which the ISO being
broken would be totally expected.