Make CDIMAGE_LOCAL_LIVEFS_ARTIFACTS more forgiving wrt artifact names
The idea behind CDIMAGE_LOCAL_LIVEFS_ARTIFACTS is that you just dump all
the files from a livefs build in a directory and point ubuntu-cdimage at
it. _Most_ projects' livefs builds name their artifacts
livecd.$PROJECT.$whatever but a few do not, as controlled by the
etc/cdimage-to-livecd-rootfs-map file. The
CDIMAGE_LOCAL_LIVEFS_ARTIFACTS code did not consult this, so trying to
use it to test one of these projects (e.g. ubuntu-core-desktop, sorry
Chris) did not work properly.

It would be possible to change the code to use the map file to look for
artifact names but as the new download_livefs_artifacts function just
splits the artifact name on "." and throws the first two parts away, we
should probably take the same approach for local artifacts as well.