introduce JH7110 SUBARCH for riscv preinstalled images
This commit goes hand in hand with the following commit:
https://git.launchpad.net/livecd-rootfs/commit/?id=949f9806469c4eaf206e9717da36fc3001fb89ab

The JH7110 is an image that supports both the milkvmars and the
visionfive2 boards.

I've not removed the other mentions of milkvmars and visionfive2
in lib/cdimage/tree.py because we still have mentions of these SUBARCHes
for older releases, so I wasn't sure if they should be removed.