do not put config values into the environment
Config objects read values from the environment which makes sense but
any changes to the config are also written to the environment which then
propagates to subprocesses. This lead to absolutely hair-tearing
problems with rebuild-requests recently and also appears to be
completely unecessary: the only subprocess we run that depends on these
environment variables is debian-cd, which is already explicitly run with
an environment based on the config.

This change is obviously a bit risky in that it's hard to be completely
sure that other subprocesses we run are not affected by any environment
variables that are being set this way. But I did do an audit of
subprocess usage:

 * bin/mirror-image-build-logs: just runs hostname, which does not care
   about any value we set in config
 * bin/rebuild-requests: is where we had the recent problems
 * bin/wait-for-package: is completely useless in the modern era and
   should be deleted
 * build.py and tree.py call lockfile, which does not care about any
   value we set
 * check_installable.py calls a subprocess but is also obsolete
 * checksums.py calls sed, which does not care about any value we set
 * germinate.py calls germinate and diff, which do not care &c
 * mail.py calls mail, which does not care &c
 * mirror.py calls ssh and apt-get which do not care &c
 * osextras.py runs some extremely obscure code to read values from the
   environment into the config, which runs before setting config values
   modifies the environment, and also wget which does not care &c
 * sign.py runs gpg, and we set GNUPG_DIR which almost looks like a
   variable gpg might care about, but isn't
 * tree.py runs zsyncmake, file, date, and mktorrent, none of which care
   &c