fix: don't overwrite older netboot artifacts
The `target_path` may point to an already existing netboot tarball,
which is likely a hardlinked copy of the previous publication. Direct
writes to this file will update the previous publication instead of
creating a new file for the current publication. Solve this by unlinking
`target_path` before writing to it.

Bug: LP:#2120484