set publish_type to "desktop" for daily-dangerous images
publish_type is set to `alternate` without this change, resulting in the
following error:

```
===== Publishing =====
Wed Aug 20 10:38:28 UTC 2025
No alternate image for amd64!
No images produced!
```

cdimage is unfortunately looking for an `alternate` image, rather than a
`desktop` image.

cdimage unfortunately looking for the wrong kind of image to publish.
With this, hopefully we have published images!