stop signing squashfs files on ISOs
As far as I know, nothing has ever verified one of these. It's possible
to imagine a world where casper in the initrd verifies the squashfs
files before mounting them, but it doesn't (for one thing, this would be
extremely slow on optical or virtual media!). If we want to go down this
route, we should back the squashfs files with dm-verity data and sign
*that* (or a checksum of that). Signing things as large as a squashfs is
a bit of an anti-pattern in general.