add a way to use lp-sign instead of gpg
The idea here is that if LP_SIGN_CONFIG is set, this overrides any GPG
settings. The code will not actually call lp-sign unless the file
pointed to by LP_SIGN_CONFIG actually _exists_ though -- so the idea is
that once LP_SIGN_CONFIG is set to a path like
$CDIMAGE_ROOT/secret/lp-sign.conf people running cdimage locally without
taking any extra steps will not get anything (SHA256SUMS, Release files
in pools) signed, which is what happens today.